# docs/swagger.yaml
basePath: /
definitions:
  basket.ResponseBasket:
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      products:
        items:
          $ref: '#/definitions/basket.ResponseProduct'
        type: array
      username:
        type: string
    type: object
  basket.ResponseProduct:
    properties:
      amount:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  login.RequestLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  pagination.Pages:
    properties:
      items: {}
      page:
        type: integer
      pageCount:
        type: integer
      size:
        type: integer
      totalCount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: yusufblt10@outlook.com
    name: Yusuf BULUT
    url: https://www.linkedin.com/in/yusufbu1ut/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Golang Bootcamp Ending Project- Basket API
  version: "1.0"
paths:
  /basket:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Takes the basket product infos and deletes it, product id is
          needed
        in: query
        name: id
        type: integer
      - description: Takes the basket product infos and deletes it, product id is
          needed
        in: body
        name: product-request
        schema:
          $ref: '#/definitions/basket.ResponseProduct'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Deletes given basket products id, it can be with query or request body
      tags:
      - Basket
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/basket.ResponseBasket'
      security:
      - ApiKeyAuth: []
      summary: Gets all basket items for the authed customer
      tags:
      - Basket
    patch:
      consumes:
      - application/json
      parameters:
      - description: Takes the basket infos and updates it, basket id and to update
          amount is needed
        in: body
        name: basket-request
        required: true
        schema:
          $ref: '#/definitions/basket.ResponseProduct'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Updates given basket item, only amount can be updated
      tags:
      - Basket
    post:
      consumes:
      - application/json
      parameters:
      - description: Takes the products and adds them
        in: body
        name: basket-request
        required: true
        schema:
          items:
            $ref: '#/definitions/basket.ResponseProduct'
          type: array
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Adds the given items to basket
      tags:
      - Basket
  /basket/complete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Completes baskets to orders with basket products
      tags:
      - Basket
  /category:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page Index
        in: query
        name: page
        type: integer
      - description: Page Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.Pages'
      summary: Gets all categories with pagination parameters page and size
      tags:
      - Category
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create categories with given csv file uploads it after that reads and
        creates (adding form-data not implemented for swagger)
      tags:
      - Category
  /login/admin:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login process takes admin' email and password. Checks the inputs
          in database and returns JWT token.
        in: body
        name: login-request
        required: true
        schema:
          $ref: '#/definitions/login.RequestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Admins login with email and password
      tags:
      - Login
...
schemes:
- http
swagger: "2.0"